requirements.txt
flask
python-dotenv
google-generativeai
gunicorn
sendgrid